pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytaxize"
dynamic = ["version"]
description = "Taxonomic toolbelt for Python"
readme = "README.md"
authors = [
    { name = "Scott Chamberlain", email = "sckott@fastmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "lxml>=4.6.0",
    "multipledispatch>=0.6.0",
    "pygbif>=0.6.0",
    "requests>=2.25.0",
]

[project.urls]
Homepage = "https://github.com/sckott/pytaxize"
Repository = "https://github.com/sckott/pytaxize.git"
Documentation = "https://pytaxize.readthedocs.io"
"Bug Tracker" = "https://github.com/sckott/pytaxize/issues"

[project.optional-dependencies]
extra = [
    "pandas>=2.3.3",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "ruff>=0.1.0",
    "sphinx",
    "sphinx-rtd-theme",
    "vcrpy>=7.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "F401", "E501"]
ignore = ["S101"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--verbose"

[tool.hatch.version]
path = "pytaxize/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["pytaxize"]
force-include = { "pytaxize/data" = "pytaxize/data" }
